exported getToggledGroupData() function to a helper file
